--- src/assets/config.ts
@@ -1,7 +1,0 @@
-export const config = {
-    pingcooldown: 60000, // 1 minute
-    creatorID: '492079026089885708', // you may not change this line!
-    hostID: '492079026089885708', // you can change this to your own ID to access hoster commands
-    logging: true, // turn to false for off, true for on
-    registercmds: true //when this option is toggled, the app will register commands globally on startup
-};
