+++ src/config.ts
@@ -1,0 +1,7 @@
+export const config = {
+    pingcooldown: 60 * 1000, // 60 seconds
+    creatorID: '492079026089885708',
+    hostID: '492079026089885708', // change this to your own ID to access hoster commands
+    logging: true,
+    registercmds: false //when this option is toggled, the app will register commands globally on startup
+};
