+++ src/slashcommands/gleaderboard.ts
@@ -1,4 +1,4 @@
-import { config } from '../assets/config';
+import { config } from '../config';
 import { EmbedBuilder } from 'discord.js';
 import { SlashCommand } from '../typings/bot';
 import { PrismaClient } from '@prisma/client'
